backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymupdf==1.25.3
python-docx==1.1.2
openai==1.70.0
python-dotenv==1.1.0
